feat(move): Move find_dict_in_list to iob_module.
