+++ lib/scripts/iob_module.py
@@ -219,3 +219,13 @@ def find_module_setup_dir(core, search_path):
     raise Exception(
         f"{iob_colors.FAIL}Setup dir of {core.name} not found in {search_path}!{iob_colors.ENDC}"
     )
+
+
+def find_dict_in_list(list_obj, name):
+    """Find an dictionary with a given name in a list of dictionaries"""
+    for i in list_obj:
+        if i["name"] == name:
+            return i
+    raise Exception(
+        f"{iob_colors.FAIL}Could not find element with name: {name}{iob_colors.ENDC}"
+    )
